Updating code: enable buttom to delete users

+++ myScript.js
@@ -27,6 +27,7 @@ function registerData() {
             showUsers();  // Actualiza la lista de usuarios después de registrar uno nuevo
         },
         error: function(xhr, status) {
+            console.error('Error al registrar usuario:', xhr.responseText);
             alert('Disculpe, existió un problema al registrar el usuario');
         }
     });
@@ -34,21 +35,19 @@ function registerData() {
 
 function showUsers() {
     $.ajax({
-        url: "http://http://ec2-54-205-126-68.compute-1.amazonaws.com:5000/registro",
+        url: "http://ec2-54-205-126-68.compute-1.amazonaws.com:5000/registro",
         type: "GET",
         contentType: "application/json",
         dataType: "json",
         success: function(users) {
             console.log(users);
-            // Limpiar la lista de usuarios antes de agregar los nuevos
             $("#user-list").empty();
             users.forEach(user => {
                 $("#user-list").append(`
                     <li class="list-group-item d-flex justify-content-between align-items-center">
                         <div>
                             <strong>${user.nombres} ${user.apellidos}</strong><br>
-                            Fecha de nacimiento: ${user.fecha_nacimiento}<br>
-                            Contraseña: ${user.password}
+                            Fecha de nacimiento: ${user.fecha_nacimiento}
                         </div>
                         <button class="btn btn-danger btn-sm" onclick="deleteUser(${user.id})">
                             Eliminar
@@ -58,12 +57,18 @@ function showUsers() {
             });
         },
         error: function(xhr, status) {
+            console.error('Error al obtener los usuarios:', xhr.responseText);
             alert('Disculpe, existió un problema al obtener los usuarios');
         }
     });
 }
 
 function deleteUser(userId) {
+    if (!userId) {
+        console.error('ID de usuario no válido');
+        return;
+    }
+    
     $.ajax({
         url: `http://ec2-54-205-126-68.compute-1.amazonaws.com:5000/usuarios/${userId}`,
         type: "DELETE",
@@ -72,7 +77,8 @@ function deleteUser(userId) {
             alert("Usuario eliminado exitosamente");
             showUsers();  // Actualiza la lista de usuarios después de eliminar uno
         },
-        error: function(xhr, status) {
+        error: function(xhr, status, error) {
+            console.error('Error al eliminar el usuario:', error);
             alert('Disculpe, existió un problema al eliminar el usuario');
         }
     });
